refactor(SinglePokemonPage): used TypeScript for component interface and props
